@@ -1,6 +1,6 @@
 module github.com/octohelm/piper
 
-go 1.25.6
+go 1.26.0
 
 tool (
 	github.com/octohelm/piper/cmd/piper
